+++ automerge-js/package.json
@@ -47,7 +47,6 @@
     "@types/uuid": "^8.3.4",
     "@typescript-eslint/eslint-plugin": "^5.25.0",
     "@typescript-eslint/parser": "^5.25.0",
-    "automerge-wasm": "^0.1.6",
     "eslint": "^8.15.0",
     "fast-sha256": "^1.3.0",
     "mocha": "^10.0.0",
@@ -56,7 +55,7 @@
     "typescript": "^4.6.4"
   },
   "dependencies": {
-    "automerge-types": "0.1.5",
+    "automerge-wasm": "0.1.7",
     "uuid": "^8.3"
   }
 }
