+++ automerge-js/test/uuid_test.ts
@@ -1,8 +1,5 @@
 import * as assert from 'assert'
 import * as Automerge from '../src'
-import * as AutomergeWASM from "automerge-wasm"
-
-Automerge.use(AutomergeWASM)
 
 const uuid = Automerge.uuid
