--- automerge-wasm/nodejs-index.js
@@ -1,5 +1,0 @@
-let wasm = require("./bindgen")
-module.exports = wasm
-module.exports.load = module.exports.loadDoc
-delete module.exports.loadDoc
-module.exports.init = () => (new Promise((resolve,reject) => { resolve(module.exports) }))
