+++ automerge-wasm/src/lib.rs
@@ -853,7 +853,7 @@ pub fn init(actor: Option<String>) -> Result<Automerge, JsValue> {
     Automerge::new(actor)
 }
 
-#[wasm_bindgen(js_name = loadDoc)]
+#[wasm_bindgen(js_name = load)]
 pub fn load(data: Uint8Array, actor: Option<String>) -> Result<Automerge, JsValue> {
     let data = data.to_vec();
     let observer = None;
